See descricption
- Made enemy attack and movement in 1 script
- Enemy and attack movement works now
- Tested on "Test towers"
