--- Assets/Scripts/Enemies/EnemyAttack.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 65b5af45ad4049641b0af97a045ea92c
